pyproject.toml
[tool.poetry]
name = "mol2ccs"
version = "0.1.1"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{ include = "mol2ccs" }]

[tool.poetry.dependencies]
python = "~3.9"
spektral = "^1.3.0"
umap-learn = "^0.5.5"
tqdm = "^4.66.1"
rdkit = "^2023.9.4"
awswrangler = "^3.4.2"
# Tensorflow Poetry and Pip all hate each other apparently. There were some problems and this comment solved them.
# https://github.com/python-poetry/poetry/issues/8271#issuecomment-1712020965
# copied from that comment
# Issue between poetry and tensorflow metadata since >=2.11
# This is a temporary workaround
# related to https://github.com/python-poetry/poetry/issues/8271
# Inspired from https://github.com/tensorflow/tensorflow/blob/adb39b04e9cb116df4659a7e2de9eea27e62f25c/tensorflow/tools/pip_package/setup.py#L148-L162
tensorflow = { version = "^2.13.0" }
tensorflow-macos = { version = "^2.13.0", platform = "darwin", markers = "platform_machine=='arm64'" }
tensorflow-intel = { version = "^2.13.0", platform = "win32" }
tensorflow-cpu = [
    { version = "^2.13.0", platform = "linux", markers = "platform_machine!='arm64' and platform_machine!='aarch64'" },
    { version = "^2.13.0", platform = "darwin", markers = "platform_machine!='arm64' and platform_machine!='aarch64'" },
]
tensorflow-cpu-aws = { version = "^2.13.0", platform = "linux", markers = "platform_machine=='arm64' or platform_machine=='aarch64'" }
# https://github.com/tensorflow/tensorflow/blob/adb39b04e9cb116df4659a7e2de9eea27e62f25c/tensorflow/tools/pip_package/setup.py#L107-L108
# https://github.com/python-poetry/poetry/issues/8271#issuecomment-1697740447
tensorflow-io-gcs-filesystem = [
    { version = ">= 0.23.1", markers = "platform_machine!='arm64' or platform_system!='Darwin'" },
    { version = "< 0.32.0", markers = "platform_system == 'Windows'" },
    { version = "<= 0.34.0", platform = "darwin", markers = "platform_machine=='arm64'" },
]
seaborn = "^0.13.1"
openpyxl = "^3.1.2"
click = "^8.1.7"
matplotlib-venn = "^0.11.10"
drugtax = "^1.0.14"
scikit-learn = "^1.4.1.post1"
scipy = "^1.12.0"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.28.0"
ruff = "^0.1.11"
pre-commit = "^3.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
